Restructure site with VizChitra 2025 recap as homepage
- Move conference content to /2025/conference
- Replace homepage with video recap grid from YouTube playlist
- Add speaker names, descriptions, and engaging hooks to video cards
- Update navigation menu to reflect new structure
- Implement uniform card layout with fixed dimensions
- Add video thumbnails and direct YouTube links
