+++ package.json
@@ -43,6 +43,7 @@
 	"dependencies": {
 		"geojson": "^0.5.0",
 		"maplibre-gl": "^5.5.0",
-		"svelte-maplibre-gl": "^0.1.8"
+		"svelte-maplibre-gl": "^0.1.8",
+		"svelte-youtube-embed": "^0.4.6"
 	}
 }
