+++ src/lib/data/youtube-videos.csv
@@ -1,15 +1,15 @@
-id,title,subtitle,thumbnail,type,code
-0nHw0TTDiEY,The Data Is Here:,But Whom Is It Serving?,https://i.ytimg.com/vi/0nHw0TTDiEY/maxresdefault.jpg,Keynote,Keynote
-htc3LwVbPgI,Data Design By Dialogue:,Visualising Data with LLMs,https://i.ytimg.com/vi/htc3LwVbPgI/maxresdefault.jpg,Keynote,Keynote
-ngjkqqjW5kY,The Eyes Have It:,Understanding Perception-Driven Data Visualisation,https://i.ytimg.com/vi/ngjkqqjW5kY/maxresdefault.jpg,Standard Talk,StandardTalk
-lNMBq8_zl_M,Data Journalism and Visualisation:,At a time of Print's decline,https://i.ytimg.com/vi/lNMBq8_zl_M/maxresdefault.jpg,Standard Talk,StandardTalk
-3dZA7iJF51g,Bringing Order to Chaos:,Designing Visualizations for Enterprise,https://i.ytimg.com/vi/3dZA7iJF51g/maxresdefault.jpg,Sponsored Talk,SponsoredTalk
-AUrJVCLiAG0,Data Visualization at Lilly:,"Designing Visuals that Solve, Simplify, and See Ahead",https://i.ytimg.com/vi/AUrJVCLiAG0/maxresdefault.jpg,Sponsored Talk,SponsoredTalk
-MmSdz7hPZGM,The invisible graph:,Leveraging AI to design accessible data visualisations,https://i.ytimg.com/vi/MmSdz7hPZGM/maxresdefault.jpg,Lightning Talk,LightningTalk
-rHigCIwrTfE,"Panel on Past,Present and Future Of Viz in Biz",,https://i.ytimg.com/vi/rHigCIwrTfE/maxresdefault.jpg,Panel Disussion,PanelDisussion
-v6u40X5HAns,The UX of Charts:,"Get the Insight Across, Every Time",https://i.ytimg.com/vi/v6u40X5HAns/maxresdefault.jpg,Lightning Talk,LightningTalk
-VqFNQPWdFdE,India's Public Health Genome:,"A Visual Story of Anomalies, Trajectories, Correlations and Geographies",https://i.ytimg.com/vi/VqFNQPWdFdE/maxresdefault.jpg,Lightning Talk,LightningTalk
-6jTHeM0KQO4,Data Visceralization:,"Moving Data, Spinning Stories",https://i.ytimg.com/vi/6jTHeM0KQO4/maxresdefault.jpg,Standard Talk,StandardTalk
-t-dm1DzU0po,Making Sense of Stories:,Structuring Qualitative Data at Scale,https://i.ytimg.com/vi/t-dm1DzU0po/maxresdefault.jpg,Lightning Talk,LightningTalk
-AGTyiQZbZCY,Democratizing Financial Data for a Growing India,,https://i.ytimg.com/vi/AGTyiQZbZCY/maxresdefault.jpg,Lightning Talk,LightningTalk
-F4Hr-DTKH70,The Unreasonable Trickiness of Tinkering,,https://i.ytimg.com/vi/F4Hr-DTKH70/maxresdefault.jpg,Lightning Talk,LightningTalk
+id,title,subtitle,thumbnail,type,code,speaker,hook
+0nHw0TTDiEY,The Data Is Here:,But Whom Is It Serving?,https://i.ytimg.com/vi/0nHw0TTDiEY/maxresdefault.jpg,Keynote,Keynote,Rukmini S,"Challenging who gets access to India's data—and who gets left out"
+htc3LwVbPgI,Data Design By Dialogue:,Visualising Data with LLMs,https://i.ytimg.com/vi/htc3LwVbPgI/maxresdefault.jpg,Keynote,Keynote,Anand S,"From bare prompt to polished chart through conversations with AI"
+ngjkqqjW5kY,The Eyes Have It:,Understanding Perception-Driven Data Visualisation,https://i.ytimg.com/vi/ngjkqqjW5kY/maxresdefault.jpg,Standard Talk,StandardTalk,Prasanta Kumar Dutta,"Where your viewer's eye goes first is no accident"
+lNMBq8_zl_M,Data Journalism and Visualisation:,At a time of Print's decline,https://i.ytimg.com/vi/lNMBq8_zl_M/maxresdefault.jpg,Standard Talk,StandardTalk,Srinivasan Ramani,"How data journalism offers print media a new lease on life"
+3dZA7iJF51g,Bringing Order to Chaos:,Designing Visualizations for Enterprise,https://i.ytimg.com/vi/3dZA7iJF51g/maxresdefault.jpg,Sponsored Talk,SponsoredTalk,Hardik Chandrahas,"Transforming enterprise chaos into clear, actionable insights"
+AUrJVCLiAG0,Data Visualization at Lilly:,"Designing Visuals that Solve, Simplify, and See Ahead",https://i.ytimg.com/vi/AUrJVCLiAG0/maxresdefault.jpg,Sponsored Talk,SponsoredTalk,Aditeya Pandey,"Empowering biopharma teams through clarity, insight and foresight"
+MmSdz7hPZGM,The invisible graph:,Leveraging AI to design accessible data visualisations,https://i.ytimg.com/vi/MmSdz7hPZGM/maxresdefault.jpg,Lightning Talk,LightningTalk,Sabhyata Jain,"Making data visualizations audible and inclusive through AI"
+rHigCIwrTfE,"Panel on Past,Present and Future Of Viz in Biz",,https://i.ytimg.com/vi/rHigCIwrTfE/maxresdefault.jpg,Panel Disussion,PanelDisussion,"Kavitha Ranganathan, Karthik Shashidharan, Aromal Ram, Rasagy Sharma","Unpacking the evolution of dataviz in business"
+v6u40X5HAns,The UX of Charts:,"Get the Insight Across, Every Time",https://i.ytimg.com/vi/v6u40X5HAns/maxresdefault.jpg,Lightning Talk,LightningTalk,Vikram Nayak,"Design data visuals that actually work and get used"
+VqFNQPWdFdE,India's Public Health Genome:,"A Visual Story of Anomalies, Trajectories, Correlations and Geographies",https://i.ytimg.com/vi/VqFNQPWdFdE/maxresdefault.jpg,Lightning Talk,LightningTalk,Saket Choudhary,"Decoding India's health patterns through data's genetic code"
+6jTHeM0KQO4,Data Visceralization:,"Moving Data, Spinning Stories",https://i.ytimg.com/vi/6jTHeM0KQO4/maxresdefault.jpg,Standard Talk,StandardTalk,Aditi Bhat,"When data becomes dance and movement becomes meaning"
+t-dm1DzU0po,Making Sense of Stories:,Structuring Qualitative Data at Scale,https://i.ytimg.com/vi/t-dm1DzU0po/maxresdefault.jpg,Lightning Talk,LightningTalk,Yashna Jhamb,"Turning messy qualitative data into meaningful change"
+AGTyiQZbZCY,Democratizing Financial Data for a Growing India,,https://i.ytimg.com/vi/AGTyiQZbZCY/maxresdefault.jpg,Lightning Talk,LightningTalk,"Rahul Beniwal, Neel Lakhwani","Making financial data engaging for India's new investors"
+F4Hr-DTKH70,The Unreasonable Trickiness of Tinkering,,https://i.ytimg.com/vi/F4Hr-DTKH70/maxresdefault.jpg,Lightning Talk,LightningTalk,Aman Bhargava,"How side projects become powerful learning playgrounds"
