+++ src/routes/+page.svelte
@@ -1,77 +1,18 @@
 <script lang="ts">
+	import { onMount } from 'svelte';
+	// Import reusable components
+	import ButtonBar from '$lib/components/ButtonBar.svelte';
+	import PolygonDivider from '$lib/components/PolygonDivider.svelte';
 	import BannerPolygon from '$lib/components/BannerPolygon.svelte';
-	import VizchitraLogo from '$lib/assets/images/viz-logo-animate.svg?raw';
-	import CallToAction from '$lib/components/CallToAction.svelte';
-	import VizChitraLogoType from '$lib/components/VizChitraLogoType.svelte';
-	import HomepageSection from '$lib/components/Homepage/HomepageSection.svelte';
 	import CustomSlantedText from '$lib/components/Common/CustomSlantedText.svelte';
-	import KeyNote from '$lib/components/LineUp/KeyNote.svelte';
-	import Speakers2025 from '$lib/components/LineUp/Speakers2025.svelte';
-	import Workshop2025 from '$lib/components/LineUp/Workshop2025.svelte';
-	import SponsorCard from '$lib/components/SponsorCards.svelte';
-	import nutanixLogo from './sponsorship/assets/nutanix-logo.png';
-	import revisualLogo from './sponsorship/assets/revisual-logo.png';	
-	import lightningChartLogo from './sponsorship/assets/lightning-chart-logo.png';
-	import hasgeekLogo from './sponsorship/assets/hasgeek-logo.png';
+	import VizchitraLogo from '$lib/assets/images/viz-logo-animate.svg?raw';
+	import Youtube from 'svelte-youtube-embed';
+	import decorBanner1 from '$lib/assets/images/decor-banner-1.png';
+	import decorBanner2 from '$lib/assets/images/decor-banner-2.png';
+	import playButton from '$lib/assets/images/play-button.png';
+	import recapThumbnail from '$lib/assets/images/recap-2025-thumbnail.png';
+	import cardDecor from '$lib/assets/images/card-decor.png';
 	import Pentagons from '$lib/components/Pentagons.svelte';
-	import Map from '$lib/components/Map.svelte';
-	import VenueCard from '$lib/components/VenueCard.svelte';
-	import FAQ from '$lib/components/FAQ.svelte';
-	import lillyLogo from './sponsorship/assets/lilly-logo.png';
-	import ltsLogo from './sponsorship/assets/lasttheoram-logo.png';
-	import straiveLogo	from './sponsorship/assets/straive-logo.png'
-
-	const sponsors = [
-		{
-			heading: 'Platinum Sponsor',
-			logo: nutanixLogo,
-			variant: 'pink' as const,
-			seed: 343442
-		},
-		{
-			heading: 'Gold Sponsor',
-			logo: lillyLogo,
-			variant: 'white' as const,
-			seed: 3434
-		},
-		{
-			heading: 'Silver & Diversity Sponsor',
-			logo: revisualLogo,
-			variant: 'yellow' as const,
-			seed: 343554354
-		},
-		{
-			heading: 'Bronze Sponsor',
-			logo: lightningChartLogo,
-			variant: 'orange' as const,
-			seed: 343554354
-		},
-		{
-			heading: 'Bronze Sponsor',
-			logo: straiveLogo,
-			variant: 'yellow' as const,
-			seed: 343554354
-		},
-		{
-			heading: 'Event Partner',
-			logo: hasgeekLogo,
-			variant: 'blue' as const,
-			seed: 342
-		},
-		{
-			heading: 'A Local Affliate for',
-			logo: 'https://images.squarespace-cdn.com/content/v1/5c6055b5fb18206d45d6b27e/1562601203642-ZHLHUGGYNWOIZHZK46U9/Data+Visualization+Society+logo+2019-05-transparent.png',
-			variant: 'green' as const,
-			seed: 34435432
-		},
-		{
-			heading: 'Media Partner',
-			logo: ltsLogo,
-			variant: 'orange' as const,
-			seed: 34435432
-		}
-	];
-
 	const whyAttend = [
 		{
 			title: 'Explore & Play',
@@ -90,295 +31,307 @@
 			backgroundColor: 'var(--color-viz-blue)'
 		}
 	];
+	// Import CSV video data
+	import videosData from '$lib/data/youtube-videos.csv';
+	import Dropdown from '$lib/components/Dropdown.svelte';
 
-	// Consolidated venue data
-	const conferenceVenue = {
-		name: 'Bangalore International Centre',
-		address: '7, 4th Main Rd, Stage 2, Domlur, Bengaluru, Karnataka 560071',
-		googleMapsURL: 'https://maps.app.goo.gl/duqbbwf2Hi3Q8dcSA',
-		transportation: [
-			{
-				method: 'Metro',
-				details: 'Indiranagar Metro Station (Purple Line) - 20 min walk'
-			},
-			{
-				method: 'Bus',
-				details: 'Buses 335-E, 333-E, 201-G, Domlur Bridge stop'
-			},
-			{
-				method: 'Auto/Cab',
-				details: 'Ask for Bangalore International Centre'
-			}
-		]
-	};
+	let isMobile = false;
 
-	const workshopVenues = [
-		{
-			name: 'Samagata Foundation',
-			address:
-				'46/1, Cobalt, 4th, Church St, Haridevpur, Shanthala Nagar, Ashok Nagar, Bengaluru, Karnataka 560001',
-			googleMapsURL: 'https://maps.app.goo.gl/4vnThEKY916vymQn6',
-			transportation: [
-				{
-					method: 'Metro',
-					details: 'MG Road Metro Station (Purple Line) - 5 min walk'
-				},
-				{
-					method: 'Bus',
-					details: 'Buses 210-G, 13, Chinnaswamy Stadium/MG Statue stop'
-				},
-				{
-					method: 'Cab',
-					details: 'Ask for Empire Restaurant on Church Street'
-				}
-			]
-		},
-		{
-			name: 'Underline Center',
-			address:
-				'3rd Floor, above Blue Tokai, #24 3rd A Cross, 1st Main Rd, Bengaluru, Karnataka 560071',
-			googleMapsURL: 'https://maps.app.goo.gl/NmHhknNLt5y2xCE89',
-			transportation: [
-				{
-					method: 'Metro',
-					details: 'Indiranagar Metro Station (Purple Line) - 20 min walk'
-				},
-				{
-					method: 'Bus',
-					details: 'Buses 335-E, 333-E, 201-G, Domlur Bridge stop'
-				},
-				{
-					method: 'Cab',
-					details: 'Ask for Bangalore International Centre'
-				}
-			]
-		}
-	];
+	// Create a unique list of 'code' values from video data
+	const uniqueCodes = [...new Set(videosData.map((video: any) => video.code))];
+
+	// Initialize objects to map video codes to corresponding CSS classes
+	const badgeClasses: Record<string, string> = {};
+	const hoverBorderClasses: Record<string, string> = {};
+
+	// Populate the class mappings dynamically
+	uniqueCodes.forEach((code: any) => {
+		badgeClasses[code] = `badge-${code.toLowerCase()}`;
+		hoverBorderClasses[code] = `card-hover-${code.toLowerCase()}`;
+	});
+
+	// Provide default fallback classes in case of missing code
+	badgeClasses['Default'] = 'badge-panel';
+	hoverBorderClasses['Default'] = 'card-hover-panel';
+
+	// Currently selected video type (filter)
+	let selectedType = 'All';
+
+	// Extract unique video types for filter buttons
+	const types = ['All', ...new Set(videosData.map((video: any) => video.type))];
+
+	// Prepare button bar data (used by <ButtonBar> component)
+	let buttonBarData = types.map((type: any, index: number) => ({
+		label: type, // Text shown on the button
+		code: type, // Internal code (used in event handling)
+		id: type + index // Unique key
+	}));
+
+	// Auto-update filtered videos whenever `selectedType` changes
+	$: filteredVideos =
+		selectedType === 'All'
+			? videosData
+			: videosData.filter((video: any) => video.type === selectedType);
+
+	/**
+	 * Opens a YouTube video in a new tab using its ID.
+	 * @param videoId - YouTube video ID
+	 */
+	function openVideo(videoId: string) {
+		window.open(`https://www.youtube.com/watch?v=${videoId}`, '_blank');
+	}
+
+	/**
+	 * Handles video type selection from the <ButtonBar> component.
+	 * Updates the selectedType reactive variable.
+	 * @param e - Event emitted by ButtonBar with `code` of the selected button
+	 */
+	function handleButtonBarChange(e: any) {
+		let activeValue = e.code;
+		selectedType = activeValue;
+	}
+
+	/**
+	 * Handles video type selection from the <Dropdown> component.
+	 * Updates the selectedType reactive variable.
+	 * @param e - Event emitted by Dropdown with `code` of the selected button
+	 */
+	function handleDropdownChange(e: any) {
+		let activeValue = e.value;
+		selectedType = activeValue;
+	}
+
+	onMount(() => {
+		const checkMobile = () => {
+			isMobile = window.innerWidth < 768;
+		};
+
+		checkMobile();
+		window.addEventListener('resize', checkMobile);
+
+		return () => {
+			window.removeEventListener('resize', checkMobile);
+		};
+	});
 </script>
 
-<div class="banner-container full-bleed relative h-[100svh]">
-	<BannerPolygon />
-	<div
-		class="logo-container pointer-events-none absolute top-1/3 left-1/2 mb-2 flex -translate-x-1/2 -translate-y-1/3 flex-col items-center rounded-md bg-white px-2 pt-2 pb-4 shadow-lg md:max-w-none md:flex-row md:gap-0 md:px-8"
-	>
-		<div class="logo">
-			{@html VizchitraLogo}
+<div class="  w-full">
+	<div class="banner-container full-bleed relative h-[80svh]">
+		<BannerPolygon />
+		<div
+			class="logo-container pointer-events-none absolute top-1/3 left-1/2 mb-2 flex -translate-x-1/2 -translate-y-1/3 flex-col items-center rounded-md bg-white px-2 pt-2 pb-4 shadow-lg md:max-w-none md:flex-row md:gap-0 md:px-8"
+		>
+			<div class="logo">
+				{@html VizchitraLogo}
+			</div>
+
+			<h3
+				class="tagline -mt-8 max-w-[20ch] border-black text-center uppercase md:mt-0 md:border-l-2 md:pl-5 md:text-left"
+			>
+				<CustomSlantedText textContent="A SPACE TO CONNECT AND CREATE WITH DATA"
+				></CustomSlantedText>
+			</h3>
+		</div>
+	</div>
+	<section class="relative h-[70svh] w-full md:h-[90svh]">
+		<!-- Top Decorative Banner -->
+		<div class="absolute top-0 z-10 w-full transition-opacity duration-300">
+			<img src={decorBanner2} alt="Decorative banner top" class="h-auto w-full object-cover" />
 		</div>
 
-		<h3
-			class="tagline -mt-8 max-w-[20ch] border-black text-center uppercase md:mt-0 md:border-l-2 md:pl-5 md:text-left"
-		>
-			<CustomSlantedText textContent="A SPACE TO CONNECT AND CREATE WITH DATA"></CustomSlantedText>
-		</h3>
+		<!-- YouTube Video Container -->
+		<div class="relative flex h-full w-full items-center justify-center bg-black">
+			<div class="relative w-full max-w-5xl md:my-36">
+				<Youtube --title-color="transparent" id="dQw4w9WgXcQ" animations={false}>
+					{#snippet thumbnail()}
+						<img
+							slot="thumbnail"
+							alt="A video on svelte"
+							src={recapThumbnail}
+							style="width: 100%; height: 100%; object-fit: contain; object-position: center;"
+						/>
+					{/snippet}
+					{#snippet play_button()}
+						<img slot="play_button" alt="A video on svelte" src={playButton} />
+					{/snippet}
+				</Youtube>
+			</div>
+		</div>
+
+		<div class="absolute bottom-0 z-4 w-full transition-opacity duration-300">
+			<img src={decorBanner1} alt="Decorative banner bottom" class=" h-36 w-full object-top" />
+		</div>
+	</section>
+	<div class="mx-auto max-w-3xl px-5 py-12">
+		<p class="mb-8 font-[Cairo] text-2xl">
+			India's first conference dedicated to <span class="text-viz-pink-dark font-bold"
+				>data visualization</span
+			>
+			brings together practitioners from
+			<span class="text-viz-orange font-bold">diverse industries and disciplines</span>
+			across the country to foster
+			<span class="text-viz-orange font-bold">learning, collaboration,</span>
+			and a shared passion for the art of
+			<span class="text-viz-pink-dark font-bold">visual data storytelling.</span>
+		</p>
+
+		<p class="mx-auto mb-6">
+			<CustomSlantedText
+				textContent="A SPACE TO CONNECT AND CREATE WITH DATA"
+				classes="text-[18px] md:text-[24px]"
+			></CustomSlantedText>
+		</p>
+
+		<p class="content-text mb-1">
+			Our goal is to build a community of diverse, interdisciplinary individuals working across the
+			visualization spectrum, and facilitate learning and connections between people from different
+			industries and disciplines who share a common interest in the power of data and storytelling.
+		</p>
 	</div>
 </div>
 
-<CallToAction />
+<div>
+	<!-- Recap Video Section with Decorative Banners -->
 
-<div class="mt-20 flex flex-col items-center gap-10">
-	<!-- OPENING SECTION -->
-	<HomepageSection>
-		<div slot="content">
-			<h1 id="intro">
-				<VizChitraLogoType></VizChitraLogoType>
-			</h1>
+	<!-- Main content container -->
+	<div class="mx-auto max-w-7xl px-4 py-8">
+		<div class="mx-auto mb-6 flex justify-center">
+			{#if !isMobile}
+				<!-- Filter Button Bar -->
+				<ButtonBar
+					data={buttonBarData}
+					activeValue={selectedType}
+					keyField="code"
+					labelField="label"
+					onSelect={handleButtonBarChange}
+				/>
+			{/if}
+			{#if isMobile}
+				<Dropdown
+					data={buttonBarData}
+					activeValue={selectedType}
+					keyField="code"
+					labelField="label"
+					onSelect={handleDropdownChange}
+				/>
+			{/if}
+		</div>
 
-			<p class="mb-8 font-[Cairo] text-2xl">
-				This first India conference brings <span class="text-viz-pink-dark font-bold"
-					>data visualization practitioners</span
+		<!-- Video Card Grid -->
+		<div class="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-3 lg:gap-6">
+			{#each filteredVideos as video (video.id)}
+				<div
+					class="relative flex h-[550px] flex-col rounded-md border border-black/20 bg-white transition-all duration-300 ease-in-out hover:border-gray-300 hover:shadow-xl"
 				>
-				from across India, representing
-				<span class="text-viz-orange font-bold">different industries and disciplines,</span>
-				together to facilitate
-				<span class="text-viz-orange font-bold">learning and connections</span>
-				around a common interest in the power of
-				<span class="text-viz-pink-dark font-bold">visual data storytelling.</span>
-			</p>
+					<div
+						class="pointer-events-none absolute inset-0 z-1 rounded-md opacity-50"
+						style="background-image: url({cardDecor}); background-repeat: no-repeat; background-position: bottom left; background-size: 320px;"
+					></div>
+					<button
+						type="button"
+						on:click={() => openVideo(video.id)}
+						class="relative z-10 flex h-full w-full cursor-pointer flex-col rounded-xl border-none bg-transparent text-left transition outline-none"
+					>
+						<!-- Card Header with Title -->
+						<div class="h-[120px] flex-shrink-0 p-4 pb-3">
+							<h3 class="text-viz-orange-dark line-clamp-4 text-xl leading-tight font-bold">
+								{video.title} <span class="text-md font-normal">{video.subtitle}</span>
+							</h3>
+						</div>
 
-			<p class="mx-auto mb-6">
+						<!-- Thumbnail Image -->
+						<div class="mx-4 mb-4 h-[200px] flex-shrink-0 overflow-hidden rounded-sm bg-black">
+							<img src={video.thumbnail} alt={video.title} class="h-full w-full object-cover" />
+						</div>
+
+						<!-- Speaker Info and Badge -->
+						<div class="flex h-[160px] flex-shrink-0 flex-col justify-between px-6 pb-6">
+							<!-- Speaker Name and Subtitle - Right Aligned -->
+							<div class="text-right">
+								<div class="text-xl leading-tight font-bold text-balance text-gray-800">
+									{video.speaker}
+								</div>
+								<p
+									class="mt-2 line-clamp-3 text-base leading-tight text-balance text-gray-600 italic"
+								>
+									{video.hook}
+								</p>
+							</div>
+						</div>
+					</button>
+				</div>
+			{/each}
+		</div>
+		<div class="mx-auto my-16 flex max-w-4xl flex-col justify-center space-y-4">
+			<p class="text-center">
 				<CustomSlantedText
 					textContent="A SPACE TO CONNECT AND CREATE WITH DATA"
 					classes="text-[18px] md:text-[24px]"
 				></CustomSlantedText>
 			</p>
-
-			<p class="content-text mb-1">
+			<p class="content-text mx-auto mb-1 text-center">
 				Our goal is to build a community of diverse, interdisciplinary individuals working across
 				the visualization spectrum, and facilitate learning and connections between people from
 				different industries and disciplines who share a common interest in the power of data and
 				storytelling.
 			</p>
-
-			<!-- <PolygonDivider></PolygonDivider> -->
-		</div>
-	</HomepageSection>
-
-	<HomepageSection>
-		<div slot="content">
-			<h2 id="speakers" class="content-heading !text-[2rem]">KEYNOTES</h2>
-		</div>
-	</HomepageSection>
-
-	<KeyNote></KeyNote>
-
-	<HomepageSection>
-		<div slot="content">
-			<h2 class="content-heading !text-[2rem]">Speakers</h2>
-
-			<p class="content-text mb-1">
-				These are the voices you won't want to miss at <strong
-					>Bangalore International Centre, Bengaluru</strong
-				>
-				on <strong>27<sup>th</sup> June</strong>
-			</p>
-		</div>
-	</HomepageSection>
-
-	<Speakers2025></Speakers2025>
-
-	<HomepageSection>
-		<div slot="content">
-			<h2 id="workshops" class="content-heading !text-[2rem]">WORKSHOPS</h2>
-			<p class="content-text mb-1">
-				So much fun to look forward to! Bear in mind, workshops are located at <b
-					>two different locations</b
-				>. Do check ahead of time.
-			</p>
-		</div>
-	</HomepageSection>
-
-	<Workshop2025></Workshop2025>
-
-	<!-- VENUES SECTION -->
-	<HomepageSection>
-		<div slot="content">
-			<h2 class="content-heading mb-6 text-center !text-[2rem]">EVENT LOCATIONS</h2>
-			<p class="content-text mb-8 text-center">
-				The conference will be held at Bangalore International Centre, while workshops are at two
-				different locations in the city.
-			</p>
-		</div>
-	</HomepageSection>
-
-	<!-- Conference Venue -->
-	<div class="w-full">
-		<h3 class="mb-6 text-center text-2xl font-bold text-gray-800">Conference Venue</h3>
-		<div class="mx-auto grid w-full grid-cols-1 gap-8 lg:grid-cols-3">
-			<div class="lg:col-span-2">
-				<div class="h-[380px] w-full overflow-hidden rounded-2xl">
-					<Map type="conference" />
-				</div>
-			</div>
-			<div class="lg:col-span-1">
-				<VenueCard venue={conferenceVenue} />
-			</div>
 		</div>
 	</div>
 
-	<!-- Workshop Venues -->
-	<div class="mb-10 w-full">
-		<h3 class="mb-6 text-center text-2xl font-bold text-gray-800">Workshop Locations</h3>
-		<div class="mx-auto grid w-full grid-cols-1 gap-8 xl:grid-cols-8">
-			{#each workshopVenues as venue, index}
-				<div class="order-2 lg:col-span-1 xl:order-none xl:col-span-2">
-					<VenueCard {venue} {index} />
-				</div>
-			{/each}
-			<div class="order-1 lg:col-span-1 xl:order-none xl:col-span-4">
-				<div class="h-full w-full overflow-hidden rounded-2xl">
-					<Map type="workshop" />
-				</div>
-			</div>
-		</div>
-	</div>
-
-	<div class="mx-auto flex max-w-4xl flex-wrap gap-4">
-		{#each whyAttend as item, i}
-			<Pentagons
-				title={item.title}
-				description={item.description}
-				backgroundColor={item.backgroundColor}
-				seed={Math.floor(Math.random() * 10)}
-			/>
-		{/each}
-	</div>
-	<h2 class="content-heading text-center !text-[2rem]">
-		Our <span class="font-bold italic">valued partners</span>
-	</h2>
-
-	<div class="mx-auto">
-		<div
-			class="ml-0 grid grid-cols-2 gap-2 pb-4 md:-ml-12 md:gap-4 md:pb-0 lg:max-w-4xl lg:grid-cols-2"
-		>
-			{#each sponsors as sponsor}
-				<SponsorCard
-					heading={sponsor.heading}
-					logo={sponsor.logo}
-					logoType="image"
-					variant={sponsor.variant}
-					seed={sponsor.seed}
+	<div>
+		<div class="mx-auto flex max-w-4xl flex-wrap gap-4">
+			{#each whyAttend as item, i}
+				<Pentagons
+					title={item.title}
+					description={item.description}
+					backgroundColor={item.backgroundColor}
+					seed={Math.floor(Math.random() * 10)}
 				/>
 			{/each}
 		</div>
 	</div>
 
-	<!-- FAQs SECTION -->
-	<HomepageSection>
-		<div slot="content">
-			<h2 class="content-heading mt-16 mb-8 text-center !text-[2rem]">
-				FREQUENTLY ASKED QUESTIONS
-			</h2>
-			<p class="content-text mb-8 text-center">
-				Got questions? We've got answers! Here are some common questions about VizChitra.
-			</p>
+	<!-- Mission Section -->
+	<div class="mx-auto max-w-4xl px-4 py-16">
+		<h2 class="mb-6 text-3xl font-bold text-gray-800">OUR MISSION</h2>
+		<p class="mb-8 text-lg text-gray-700">
+			To foster a vibrant <span class="font-bold">community of data storytellers in India</span>,
+			bridging technical analysis and design expertise to shape perspectives and drive change
+		</p>
+
+		<div class="mb-8 text-center">
+			<CustomSlantedText
+				textContent="AN INDIAN DATA VISUALIZATION COMMUNITY"
+				classes="text-[18px] md:text-[24px]"
+			></CustomSlantedText>
 		</div>
-	</HomepageSection>
+
+		<p class="mb-8 text-lg text-gray-700">
+			The VizChitra community plans to drive this using these three main pillars of work:
+		</p>
+
+		<div class="space-y-6">
+			<div>
+				<h3 class="mb-2 text-xl font-bold text-gray-800">1. Consider and Curate:</h3>
+				<p class="text-gray-700">
+					Build a rhythm of curated events to spread the practice of data visualization
+				</p>
+			</div>
+
+			<div>
+				<h3 class="mb-2 text-xl font-bold text-gray-800">2. Cultivate and Care:</h3>
+				<p class="text-gray-700">
+					Nurture a fertile space for learning and sharing of data visualization skills
+				</p>
+			</div>
+
+			<div>
+				<h3 class="mb-2 text-xl font-bold text-gray-800">3. Create and Collaborate:</h3>
+				<p class="text-gray-700">
+					Express and co-create to push the boundaries of data visualization
+				</p>
+			</div>
+		</div>
+		<PolygonDivider></PolygonDivider>
+	</div>
 </div>
-<div class=" w-full">
-	<FAQ />
-</div>
-
-<style>
-	.banner-container {
-		width: 100vw;
-		position: relative;
-	}
-
-	.banner-container::before,
-	.banner-container::after {
-		content: '';
-		position: absolute;
-		left: 0;
-		right: 0;
-		height: 25vh;
-		pointer-events: none;
-		z-index: 1;
-	}
-
-	@media (min-width: 768px) {
-		.banner-container::before,
-		.banner-container::after {
-			height: 40vh;
-		}
-	}
-
-	.banner-container::before {
-		top: 0;
-		background: linear-gradient(
-			to bottom,
-			rgb(255 255 255) 0%,
-			rgba(255, 255, 255, 0.8) 5%,
-			transparent 100%
-		);
-	}
-
-	.banner-container::after {
-		bottom: 0;
-		background: linear-gradient(
-			to top,
-			rgb(255 255 255) 0%,
-			rgba(255, 255, 255, 0.8) 5%,
-			transparent 100%
-		);
-	}
-</style>
