--- tina/.gitignore
@@ -1,1 +1,0 @@
-__generated__
